Added Roads and MapGraph
Adds roads as a type of building as well as giving buildings a bool value to tell if it's a road. Also added MapGraph which physically represents whether buildings are connected on the map. Started implementation on the BuildingMap, a weighted graph, which shows the connections directly between the buildings, need to implement the connections still.