Changed satisfaction to include building distance
Changed satisfaction to an increasing system instead of decay, it uses building distances from accommodation to calculate the multiplier as well as a minor increase for each building placed. Also fixed some issues with pausing not affecting parts of code that it should. And lowered the UI scaling for the satisfaction score to 10x instead of 1000.